requirements.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.7
psycopg2==2.9.3
python-decouple==3.6
